python/dropbox/requirements.txt
dropbox
python-dotenv

